feat: 유스케이스 프로토콜 생성
