+++ Acha/Acha/Domain/Entity/Auth/LogInData.swift
@@ -1,0 +1,13 @@
+//
+//  LogInData.swift
+//  Acha
+//
+//  Created by hong on 2022/11/27.
+//
+
+import Foundation
+
+struct LoginData: Equatable {
+    let email: String
+    let password: String
+}
