+++ Acha/Acha/Presentation/Auth/UseCase/SignUpUsecase.swift
@@ -9,6 +9,19 @@ import Foundation
 import RxSwift
 import RxRelay
 
+protocol SignUpUsecaseProtocol: EmailValidatable, PasswordValidatable, NickNameValidatable {
+    var emailRelay: BehaviorRelay<String> {get set}
+    var passwordRelay: BehaviorRelay<String> {get set}
+    var nickNameRelay: BehaviorRelay<String> {get set}
+    
+    var emailValidation: Bool {get set}
+    var passwordValidation: Bool {get set}
+    var nickNameValidation: Bool {get set}
+    
+    func isSignAble() -> Bool
+    func signUp() -> Observable<String>
+}
+
 final class SignUpUsecase {
     
     var emailValidation: Bool = false
